Code to compute day3 part 1.
This might be the hard way of doing it. There feels like there should be
an easy way, but I was not able to come up with it quickly.

We can figure out what each of the "rings" numbers should be. Only
odd-numbered elements per edge are considered a full "ring". With the
squareroot of the input giving us which ring it's in, we can compute
what each side of the square is.

Then, the Manhattan distance is the total number of steps away from the
center.  It's one minus the current ring number, plus however many
horizontal steps it is off the center.

Like I said, probably an easier way to do this.